Convert the 'status' attribute to the column in database
The cache library ('whtht/perfectly-cache') that is requested to be used performs
caching according to the SQL query. Previously, the status attribute was
determined at that moment based on the current date and time. This caused the
generated sql query to constantly change and the caching library to be unable to
do its job.

- Removed 'status' attribute from the Article model.
- Added 'status' enum column to the database.
- 'getRoleFromString' method renamed to 'getEnumFromString' and moved to
  theabstract class 'BaseEnum' to make it reusable.
- Migrated from 'getRoleFromString' method to 'getEnumFromString' method.
- Added 'doctrine\dbal' to use enum type in the testing sqlite database.
- Added ArticleStatus enum class and created unit test case for it.